src/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kafka-python
pydantic==2.5.0
